Snapshot sender app updates files sent counts, sent status.
To enable files to be sent to crown as they are produced (rather than
when the entire collection has been exported) snapshot sender needs
to be able to set the 'Sent' status on the status table and it
needs to increment the 'FilesSent' value on the status table.

When the 'FilesSent' equals the 'FilesExported' and the status is 'Exported'
then snapshot sender knows that the collection is finished and it can send
the success indicator and set the status to 'Sent'.