SEND SUCCESS ON NO FILES EXPORTED (#79)
If HTME legitimately doesn't export any files for a collection (for instance if
it's an incremental run and there were no changes that day) it will send a
message indicating that this was the case. On receipt of such a message snapshot
sender marks the collection as 'Sent' in dynamodb.

Fixed the integration tests which had been failing in circle.

Updated the unit test to ensure the headers on the NiFi request work for all
filename types.

Sorted the integration test containers into their own sub-directories.

Have circle use make to run the tests so that a successfull local run is a
good indicator of a to-be-successful run in circle.

Remove redundant files.

Intellij code inspection updates.