Additional headers in the NiFi post.
To allow NiFi to request a resend (of a file that it failed to process
for example) it has to be able to send all the information that
snapshot-sender received from HTME, this information is placed in
the headers of the POST request and the 2 that were missing were
the 'reprocess files' flag and the 'shutdown' flag - so include these
in the headers.

Changed a couple of Makefile targets to read more naturally e.g.
'make jar' instead of 'make build-jar'.