Send success file on request.
Snapshot sender no longer sends a success file when it has determined
that the collection has been sent in full. This is because files can
fail downstream after they have been accepted by nifi. So,
snapshot-sender will set the status to 'Sent' when it has sent all
the files but not send the success indicator. When all the
files have landed in crown snapshot-sender will be invoked in a mode
whereby it sends the success indicator and sets the status to
'Success' on dynamodb.