Additional headers in the NiFi post. (#80)
* Additional headers in the NiFi post.

To allow NiFi to request a resend (of a file that it failed to process
for example) it has to be able to send all the information that
snapshot-sender received from HTME, this information is placed in
the headers of the POST request and the 2 that were missing were
the 'reprocess files' flag and the 'shutdown' flag - so include these
in the headers.

Changed a couple of Makefile targets to read more naturally e.g.
'make jar' instead of 'make build-jar'.

* De-odorized.

* De-odorized part 2.

* sync

* Send success file on request.

Snapshot sender no longer sends a success file when it has determined
that the collection has been sent in full. This is because files can
fail downstream after they have been accepted by nifi. So,
snapshot-sender will set the status to 'Sent' when it has sent all
the files but not send the success indicator. When all the
files have landed in crown snapshot-sender will be invoked in a mode
whereby it sends the success indicator and sets the status to
'Success' on dynamodb.

* SonarCloud code duplication updates.

* Unit test fix.