Set status updates.
If snapshot sender receives a message from HTME indicating that no
files were exported it sets the status to 'Received' and send the
success indicator.

On receiving the 'send a success indicator' message is does that
and leaves the status as-is.

On determining that all exported files have been sent it sets the
status to 'Sent' but does not send a success indicator.